+++ include/quill/sinks/FileSink.h
@@ -20,6 +20,7 @@
 #include <cstdio>
 #include <cstring>
 #include <ctime>
+#include <memory>
 #include <string>
 #include <utility>
 
@@ -97,6 +98,11 @@ public:
    * The default value is false.
    * @param value True to perform fsync, false otherwise.
    */
+  QUILL_ATTRIBUTE_COLD void set_fsync_enabled(bool value) { _fsync_enabled = value; }
+
+  [[deprecated(
+    "This function is deprecated and will be removed in the next version. Use set_fsync_enabled() "
+    "instead.")]]
   QUILL_ATTRIBUTE_COLD void set_do_fsync(bool value) { _fsync_enabled = value; }
 
   /**
@@ -106,6 +112,46 @@ public:
    */
   QUILL_ATTRIBUTE_COLD void set_open_mode(char open_mode) { _open_mode = open_mode; }
 
+  /**
+   * @brief Sets the user-defined buffer size for fwrite operations.
+   *
+   * This function allows you to specify a custom buffer size for fwrite, improving efficiency
+   * for file write operations.
+   *
+   * To disable custom buffering and revert to the default size, pass a value of 0.
+   *
+   * @note By default, a buffer size of 64 KB is used.
+   * @param value Size of the buffer in bytes. If set to 0, the default buffer size will be used.
+   */
+  QUILL_ATTRIBUTE_COLD void set_write_buffer_size(size_t value)
+  {
+    _write_buffer_size = (value == 0) ? 0 : ((value < 4096) ? 4096 : value);
+  }
+
+  /**
+   * Sets the minimum interval between `fsync` calls. This specifies the minimum time between
+   * consecutive `fsync` operations but does not guarantee that `fsync` will be called exactly
+   * at that interval.
+   *
+   * For example, if some messages are flushed to the log and `fsync` is skipped because it
+   * was previously called, and no further messages are written to the file, `fsync` will not
+   * be called even if the minimum interval has passed. This is because the previous call was
+   * skipped due to the interval, and no new messages necessitate another `fsync` call.
+   *
+   * This feature is intended to mitigate concerns about frequent `fsync` calls potentially causing
+   * disk wear.
+   *
+   * Note: This option is only applicable when `fsync` is enabled. By default, the value is 0,
+   * which means that `fsync` will be called periodically by the backend worker thread when
+   * messages are written to the file, irrespective of the interval.
+   *
+   * @param value The minimum interval, in milliseconds, between `fsync` calls.
+   */
+  QUILL_ATTRIBUTE_COLD void set_minimum_fsync_interval(std::chrono::milliseconds value)
+  {
+    _minimum_fsync_interval = value;
+  }
+
   /** Getters **/
   QUILL_NODISCARD bool fsync_enabled() const noexcept { return _fsync_enabled; }
   QUILL_NODISCARD Timezone timezone() const noexcept { return _time_zone; }
@@ -114,9 +160,16 @@ public:
     return _filename_append_option;
   }
   QUILL_NODISCARD std::string const& open_mode() const noexcept { return _open_mode; }
+  QUILL_NODISCARD size_t write_buffer_size() const noexcept { return _write_buffer_size; }
+  QUILL_NODISCARD std::chrono::milliseconds minimum_fsync_interval() const noexcept
+  {
+    return _minimum_fsync_interval;
+  }
 
 private:
   std::string _open_mode{'w'};
+  size_t _write_buffer_size{64 * 1024}; // Default size 64k
+  std::chrono::milliseconds _minimum_fsync_interval{0};
   Timezone _time_zone{Timezone::LocalTime};
   FilenameAppendOption _filename_append_option{FilenameAppendOption::None};
   bool _fsync_enabled{false};
@@ -145,6 +198,12 @@ public:
                  nullptr, std::move(file_event_notifier)),
       _config(config)
   {
+    if (!_config.fsync_enabled() && (_config.minimum_fsync_interval().count() != 0))
+    {
+      QUILL_THROW(
+        QuillError{"Cannot set a non-zero minimum fsync interval when fsync is disabled."});
+    }
+
     if (do_fopen)
     {
       open_file(_filename, _config.open_mode());
@@ -275,7 +334,15 @@ protected:
                              " errno: " + std::to_string(errno) + " error: " + strerror(errno)});
     }
 
-    assert(_file && "open_file always returns a valid pointer or throws");
+    if (_config.write_buffer_size() != 0)
+    {
+      _write_buffer = std::make_unique<char[]>(_config.write_buffer_size());
+
+      if (setvbuf(_file, _write_buffer.get(), _IOFBF, _config.write_buffer_size()) != 0)
+      {
+        QUILL_THROW(QuillError{std::string{"setvbuf failed error: "} + strerror(errno)});
+      }
+    }
 
     if (_file_event_notifier.after_open)
     {
@@ -310,14 +377,22 @@ protected:
   /**
    * Fsync the file descriptor.
    * @param fd File descriptor.
-   * @return True if successful, false otherwise.
    */
-  static bool fsync_file(FILE* fd) noexcept
+  void fsync_file(FILE* fd) noexcept
   {
+    auto const now = std::chrono::steady_clock::now();
+
+    if ((now - _last_fsync_timestamp) < _config.minimum_fsync_interval())
+    {
+      return;
+    }
+
+    _last_fsync_timestamp = now;
+
 #ifdef _WIN32
-    return FlushFileBuffers(reinterpret_cast<HANDLE>(_get_osfhandle(_fileno(fd)))) != 0;
+    FlushFileBuffers(reinterpret_cast<HANDLE>(_get_osfhandle(_fileno(fd))));
 #else
-    return ::fsync(fileno(fd)) == 0;
+    ::fsync(fileno(fd));
 #endif
   }
 
@@ -353,6 +428,8 @@ private:
 
 private:
   FileSinkConfig _config;
+  std::chrono::steady_clock::time_point _last_fsync_timestamp{};
+  std::unique_ptr<char[]> _write_buffer;
 };
 
 QUILL_END_NAMESPACE